+++ tests/testthat/test-collect.R
@@ -1,5 +1,5 @@
 test_that("axes and titles are collected correctly for multi-cell plots", {
-  plots  <- wrap_plots(rep(list(p1), 8))
+  plots <- wrap_plots(rep(list(p1), 8))
   layout <- plot_layout(
     design = "12345\n62378",
     axes = "collect",
@@ -15,7 +15,6 @@ test_that("axes and titles are collected correctly for multi-cell plots", {
 })
 
 test_that("axis columns are properly resized", {
-
   p5 <- p1 + scale_y_continuous(
     labels = function(x) paste0("a long axis label signifying ", x)
   )
@@ -40,8 +39,28 @@ test_that("axis titles are collected across empty areas", {
       axis_titles = "collect",
       design = "#AB\nC#D\nEF#"
     )
-    expect_doppelganger(
-      "Empty areas doesn't interfere with title collection",
-      plots
+  expect_doppelganger(
+    "Empty areas doesn't interfere with title collection",
+    plots
+  )
+})
+
+test_that("collect guides works well", {
+  expect_doppelganger(
+    "collect normal guides",
+    wrap_plots(p1 + p3, guides = "collect")
+  )
+  p_guides <- p3 + scale_color_continuous(guide = guide_colorbar(
+    theme = theme(legend.key.height = unit(1, "null"))
+  ))
+  expect_doppelganger(
+    "collect guides with null unit",
+    wrap_plots(p1 + p_guides, guides = "collect")
+  )
+  expect_doppelganger(
+    "collect guides with multiple plots with null unit",
+    wrap_plots(p1 + p_guides + p_guides + labs(color = "another"),
+      guides = "collect"
     )
+  )
 })